Update OAuth functional tests to match profile
Now that this endpoint supports both authenticated and unauthenticated
access, we can no longer use the status code of the response to
`/api/profile` to test whether the token is valid.

This may be a temporary fix, as we're looking into a piece of work to
unify how we use HTTP status codes to handle missing and invalid tokens
across our API, alongside an endpoint to return information about the
current token, but for now this is the best way to determine whether a
token provides access on behalf of a user.