Add fetch method to auth token service
This will return a database auth token (`h.models.Token`) rather than a
`h.auth.tokens.Token`, for when we need to access more data than the
userid and the expiration time.

Note that this method does not cache the result of the database query,
thus potentially running the query multiple times during a request.