Make our JWT timestamp type validation stricter
The `pyjwt` library is pretty permissive, allowing `None` values and
strings of digits in its timestamps. This causes us a bunch more work
when trying to do calculations around the life of the token. Given that
RFC 7519 insists that these values MUST be numbers[1], we can reasonably
reject values that are not number types.

[1]: https://tools.ietf.org/html/rfc7519#section-4.1.4 and onwards.