Add `h.client_url` setting to specify where to load client boot script from
Add a setting which specifies the URL for the client's boot script.
Client assets are expected to be loaded from the same URL, but with a
trailing slash added.

The latest stable production client is loaded by default. It can be
overridden in local development by setting the `CLIENT_URL` env var.