Factor out a GrantToken class
This shouldn't change any of the behaviour of the token validation code,
but splits apart the concerns of:

  * parsing the JWT, checking it has the necessary claims in the right
    format, and is valid at the current time
  * interacting with the session to get hold of the `AuthClient` and
    user

This should make it easier to add more checks to the token -- for
instance, checking that the effective duration of the token doesn't
exceed a certain threshold. It also frees up the `OAuthService` from
dealing with the details of how the `pyjwt` library works (for instance,
that we have to use the same `decode` method for the initial parsing and
for checking whether the token is valid).