Add a docker-compose.yml for services
While running the h development server itself inside Docker using
Compose is a bit fiddly and possibly not what we want to do, using
Compose to coordinate the backing services seems to be straightforward
and helpful.

Add a docker-compose.yml which configures the Postgres, Elasticsearch,
and RabbitMQ containers, and update the dev documentation to use it.