Enforce a maximum lifetime for grant tokens
To mitigate the security risks if a grant token is compromised, we're
now making sure that these tokens will only be honoured if they are
short-lived -- namely, that they have a start time (`nbf`) and an expiry
time (`exp`) and that the difference between them is no more than 10
minutes.

As part of this, the format validation on these timestamps is now more
restrictive. The `pyjwt` library is pretty permissive, allowing `None`
values and timestamp values as strings (e.g. `"1487339382"`).  Given
that RFC 7519 insists that these values MUST be numbers[1], we can
reasonably reject values that are not number types.

[1]: https://tools.ietf.org/html/rfc7519#section-4.1.4 and onwards.