Split error handling for missing/invalid tokens
The fact that the `GrantToken` class was raising exceptions that
mentioned the "assertion parameter" was a little strange, as the class
isn't aware that's where the token has come from. This is now split into
handling a missing `assertion` parameter, and handling a value for that
parameter that isn't parseable as a JWT.

Hat tip to @chdorner for pointing this out.