Update key derivation salt to be a random value
Ideally the HKDF salt should be a value as long as the underlying hash
function output and random. Historically we'd used the fixed string
'h.security' for this, but stronger security guarantees are offered by
using a random salt.

This commit also removes the ability to configure algorithm and output
length when deriving keys, and allows the salt to be overridden in an
environment variable.