Unify grant token error messages
As noted by @nickstenning[1], our error messages are quite inconsistent,
sometimes referring to a "grant token", sometimes to a "JWT" and
sometimes to a "JWT token"[2]. Now everything refers to a "grant token".

I've taken the opportunity to make the error messages a bit more
consistent in other respects as well, so that they now include both a
human-friendly description of a claim (e.g. 'subject') and the short
version in the token (e.g. 'iss'). I've made an exception where this
would make the message more difficult to understand -- for instance,
"grant token expiry time (exp) is in the past" would be harder to
understand than "grant token is expired".

[1]: https://github.com/hypothesis/h/pull/4381#issuecomment-281071051
[2]: https://en.wikipedia.org/wiki/RAS_syndrome