Refactor grant token errors to remove duplication
Now that we have a common style of error message for certain error
cases, we can pull those cases out into internal helper methods on the
`GrantToken` class and remove a lot of the repetition.

There's still some duplication in the `OAuthService` itself, which may
be a sign that the boundary between these two classes could be cleaner,
but I think this is still an improvement.