Fix generated RangeSelector objects
The RangeSelector objects generated by the client and stored in the
database are not valid RangeSelectors as defined by the W3C Annotation
Model specification.

Specifically, the RangeSelector in the spec is a generic range which has
`startSelector` and `endSelector` properties, whereas the client's
"RangeSelector" only ever contains an XPath + offset for start and end.

Add support for remapping the old-style RangeSelector in the JSONLD
presenter, to bring us inline with the published spec.