Match dot- and case-variants when looking up by username (#4397)
We've had a number of support tickets from users who have created their
accounts with usernames such as "JoeBloggs", and are later confused
because they can't log in as "joebloggs".

This gets particularly confusing when, having had the login form tell
them "User does not exist," they go and try and sign up again, at which
point the signup form tells them that the username is taken.

This commit addresses this at the database and ORM level, by ensuring
that all for comparisons against the username column, both sides of the
comparison are normalised first. For example

    request.db.query(models.User).filter_by(username='Joe.Bloggs')

which would previously have resulted in a WHERE clause of

    WHERE username = 'Joe.Bloggs'

will now result in a WHERE clause more like

    WHERE lower(replace(username,     '.', '')) =
          lower(replace('Joe.Bloggs', '.', ''))

In order to make this feasible, I've made some changes to the user
model. The separate `uid` column is gone, replaced by a unique
functional index on the tuple:

    lower(replace(username, '.', '')), authority

This simultaneously achieves three ends:

1. It makes lookups by normalised username fast.
2. It ensures uniqueness of (normalised username, authority)
3. It (together with some changes to the UserIDComparator) makes lookups
   by userid fast.

With this in place, we're also able to remove the index on `authority`
alone.