Fix userid index in database
Unfortunately, the migration in c9a5aeb12 was incorrect, as it failed to
specify that the index to be added was unique.

This migration fixes the problem by moving the old index out of the way,
recreating it as unique, and then dropping the old index.

Note that the index is still referred to using a normal index prefix
("ix__") rather than a unique constraint prefix ("uq__") because unique
constraints cannot be created using functional indices.