Rename 'client_url' to 'embed_url'
Change the name to reflect the fact that this var contains the URL of
the /embed.js script rather than the client URL directly.

We _could_ now change app.html to include the client URL directly, but
I'll do this as part of a separate PR since it is a functional change.