Add Referrer-Policy and X-XSS-Protection headers
This commit adds a Referrer-Policy[1] and X-XSS-Protection[2] header to
every response.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referrer-Policy
[2]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-XSS-Protection

Referrer-Policy aims to address problems of URL leakage -- users
visiting HTTPS web pages expect the URLs they visit to be protected,
which they're not unless the Referer information is appropriately
redacted. A value of "strict-origin-when-cross-origin" means that:

1. pages on the same origin will receive a full Referer header

2. pages on a different secure origin will receive an origin-only
   Referer header (i.e. the URL path is redacted)

3. pages on a different insecure origin will receive no Referer header.

This commit also adds a strict X-XSS-Protection header. This will be
largely unnecessary when we deploy a strict CSP policy, but can
nonetheless protect users from reflected XSS in older (pre-CSP)
browsers.