Clear groupfinder cache on transaction end
For the same reasons why we're clearing the user cache on transaction
end.

We previously cleared the groupfinder service cache on `after_commit` and
`after_rollback` SQLAlchemy events. This caused issues where the view
was raising an exception, because SQLAlchemy only fires the
`after_rollback` event for an explicit rollback.

pyramid_tm and zope.sqlalchemy will close the database transaction when
the view raises an exception. A call to `SessionTransaction.close` does
implicitly send a `ROLLBACK` SQL query to the database, but it doesn't
emit the `after_rollback` event that we're relying on in order to clear
out the user service cache.
This results in a exception views potentially accessing detached
database objects which in turn throws another exception, which then
returns a generic 500 error back to the user.