Opt /app.html views out of Content-Security-Policy protections
There are a lot of CSP exceptions being generated from this page, many
of which appear to be the result of adware or misbehaving browser
extensions.

It may well be a good idea to have CSP enabled for the client, but in
order to see more clearly if there are real CSP issues on the rest of
the site, and to avoid breaking the client, we must temporarily suppress
CSP headers for these views.