Clean up the header construction for CSP view deriver
The code responsible for constructing the CSP header value was somewhat
inscrutable, apparently to account for the possibility that null values
would be provided in the settings.

Cleaning up the default CSP policy settings allows us to remove this
complexity and make it a bit more obvious how the header value is
constructed.