Slim down the healthcheck endpoint
Having the healthcheck/status endpoint talk to all the backing services
is not the greatest idea in the world, on reflection.

The problem (apart from the fact that it's quite slow) is that it
tightly couples the health of the service to the health of the backing
services. If ElasticSearch is temporarily unavailable, a failing
healthcheck will take the entire site offline as all instances are
removed from the load balancer. By contrast, a healthcheck that returns
only the health of the web application will allow those parts of the
site that don't depend on ElasticSearch to keep working.

The counter-argument is that if a single instance is having networking
issues, it is useful to have it removed from the load balancer so it
doesn't keep serving requests.

This commit slims down the healthcheck endpoint to only check Postgres.
Very little of the site currently works without a database so this seems
like a reasonable compromise for now.