Add intersphinx mapping to client repo
These means the :ref:`...` cross-references (and other kinds of
cross-references, like :js:func:, :option:, :envvar:, ...) can refer to
things defined in the client repo's docs, and Sphinx will generate URLs
to the client docs.

There aren't actually and refs in these docs that make use of this yet
but it's good to have it here.