Update README now that client has own repo
Update the README to clarify that this is the git repo for the web
service, and link to the client's separate repo.

I also removed the mention of Chrome extension and Firefox bookmarklet -
there is just a link to the client described as a "browser-based
annotator" and the client's own repo then links to the Chrome extension.