Rename Sphinx project to "h"
Rename the Sphinx project from "The Hypothesis Annotation Framework
Documentation" to just "h", which is the name of the web app that runs
most of https://hypothes.is. This is the documentation for the web
app only, the client (and browser extensions etc) have their own
separate documentations. And I'm not sure there was ever really an
annotation framework.