Add an ADR summarising our API design approach
Following on from a recent team discussion†, we ended up agreeing on a
set of conventions for our API, specifically around which principles of
RESTful architecture we wanted to follow, and how to deal with
interdependent resources. This seemed like a consensus worth
documenting, for the sake of making future API design easier and
quicker, and hopefully making future API functionality look consistent
for those hoping to integrate with us.

† This discussion took place in the Hypothesis' backend Slack channel,
  between 2017-03-22 at 10:56 UTC[1], and continuing to roughly 09:50
  UTC the following day[2]. Slack doesn't really have the URL scheme
  necessary to link to conversation ranges.

[1]: https://hypothes-is.slack.com/archives/C1MA4E9B9/p1490180173520032
[2]: https://hypothes-is.slack.com/archives/C1MA4E9B9/p1490262483710904