Add formatter layer and integrate it into JSON annotation presentation
Formatters have to implement the new
`h.formatters.interfaces.IAnnotationFormatter` interface and should then
be added to the list of formatters inside the AnnotationJSONPresentation
service. These formatters will need to return a dictionary object for
an annotation which is getting merged into the JSON dictionary for the
annotation, effectively allowing a better separation and modular
approach to how we construct a JSON representation of an annotation.