Move memex.search to h.search (#4448)
This is part of a project to reintegrate the `memex` package into the
main `h` package: https://notes.wtk.io/2017/03/08/reintegrating-memex

Note that this commit does not remove all reference to the "memex" name
from this package: the configuration directives are still called
"memex_*". We can address these in subsequent commits.