Dynamically resolve upstream hostnames in nginx
Unfortunately, nginx does not respect DNS TTLs for the hostnames of
upstream servers by default. This means that if the DNS record for
FALLBACK_UPSTREAM is updated, the nginx instance will continue to send
requests to old hosts. This is a very real possibility if
FALLBACK_UPSTREAM points to an Elastic Load Balancer, where hosts may be
dynamically swapped out to ensure continued availability of the load
balancer.

This can be fixed by moving the upstream hostname into a variable, per
the documentation for proxy_pass[1]:

   A server name, its port and the passed URI can also be specified using variables:

       proxy_pass http://$host$uri;

   or even like this:

       proxy_pass $request;

   In this case, the server name is searched among the described server
   groups, and, if not found, is determined using a resolver.

This also means that we have to specify a valid "resolver" directive in
the nginx configuration. This is done by parsing /etc/resolv.conf within
the container and injecting the appropriate directive.

In order to make it easier to understand what's going on when templating
the nginx configuration, this commit also switches to using the envsubst
program[2] from gettext rather than arcane sed invocations.

[1]: https://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_pass
[2]: https://www.gnu.org/software/gettext/manual/html_node/envsubst-Invocation.html