Add lru_cache_in_transaction helper
An increasing number of services are gaining features to cache a
particular dataset for the duration of the database session.

Rather than have each of these services to implement and test their own
caching behaviour, it makes sense to introduce a general purpose tool to
achieve the same ends.

This commit introduces a decorator, `lru_cache_in_transaction`, which is
a thin wrapper around the `lru_cache` decorator from Python >= 3.3, with
the added bonus that it clears its internal cache when the database
transaction ends.