Ensure h.models is imported first in the WebSocket server
Unfortunately, the previous commit (f19ff6390) exposed a latent bug in
the WebSocket server configuration, where `h.models` was not guaranteed
to be included before any models from `memex`. See the note at the top
of `h.models` for an explanation of why this is a problem.

The underlying issue will go away very soon when we move `memex.models`
back into the `h` package, but for now the fix is to move the inclusion
of `h.models` to the top of the app bootstrap, just as in `h.app`.