Rename request.authenticated_user to request.user
The additional verbosity is cumbersome and not particularly
enlightening. We originally had to use "authenticated_user" rather than
"user" because the latter was already used by code that has since been
removed.

The "authenticated_" prefix mirrors the Pyramid-provided
"request.authenticated_userid", but that property is so named to
distinguish it from "request.unauthenticated_userid", whereas there is
no such corresponding property for us to distinguish ourselves from with
this one.