Rename request.auth_domain to request.authority
We are now consistently using the term "authority" to refer to the
domain-like namespace that we use to partition users, groups, and
annotations within our database.

"request.auth_domain" is a property which predates the use of the term
"authority" within the codebase, but essentially represents the same
concept, and so it makes sense to unify the language here.