Emit a warning if the deprecated AUTH_DOMAIN env var is used
This commit adds back support for reading the value of the h.authority
setting from the AUTH_DOMAIN environment variable, in order to maintain
backwards compatibility for existing installations.

It also adds a DeprecatedSetting helper to make it easier to print
deprecation warnings to the log when deprecated usage is detected.