Move groupfinder (IGroupService) service from memex to h
This commit moves the interface definition for the "groupfinder service"
(AKA memex.interfaces.IGroupService) from memex into h. Doing this
allows us to:

- move the registration of the groupfinder service from h.app back into
  h.services (as we no longer need to explicitly trigger Pyramid's
  configuration conflict detection machinery)
- remove the dummy implementation of IGroupService (memex.groups)
  entirely

In future, it may make sense to consolidate h.services.group and
h.services.groupfinder, as suggested by a "FIXME" comment in the latter.
In addition, the indirection provided by the IGroupService interface
itself may no longer be justified. To keep things simple, this commit
doesn't tweak how this code is referenced, and sticks to just changing
where it lives.

This is part of a project to reintegrate the `memex` package into the
main `h` package: https://notes.wtk.io/2017/03/08/reintegrating-memex.