Add a UserPasswordService for managing user passwords (#4468)
This commit adds a service which has responsibility for checking (and
updating) user passwords. This is in preparation for removing that
functionality from the user model.

This is desirable primarily because it makes it much easier to make our
tests fast. At the moment, many tests which interact with User model
instances end up computing the 12-round bcrypt hash of passwords, which
is *very* slow.

By encapsulating the password checking functionality in a service of its
own we can limit the performance impact of password checking in tests to
just this service.