Use the UserPasswordService when checking passwords (#4469)
This commit makes all parts of the code that *check* user passwords do
so using the UserPasswordService. Note that it does not change the parts
of the code that *update* user passwords, and so we are relying on the
fact that passwords set by the code in the User model will be correctly
checked by the code in the new service.

Previously, some parts of the code checked user passwords directly using
`User.check_password`, while at login, passwords were checked by the
UserService's `login` method.

This commit renames the UserService's `login` method to
`fetch_for_login` and removes the part of that method that checks user
passwords.

Then, we:

- update all parts of the code which called `User.check_password`
  directly to use the user password service instead

- update all parts of the code which called the UserService's `login`
  method to use `fetch_for_login` and separately check the password with
  the user password service.

N.B. There remain a couple of calls to `User.check_password` in the
tests, but these are used to check that *setting* passwords is happening
correctly, and will be addressed in a later commit.