Mark moderated annotations as nipsa'd in search index
For now, this is done the same way we already check if the user is
nipsa'd, and then add a `"nipsa": true` item to the search index
payload.

Ideally, we would also get rid of the `AnnotationTransformEvent` system
and build on top of the recently introduced formatters (or something
similar to that). But since the card for annotation moderation is
already quite big I've opted for extending the current system instead.