Document CLIENT_URL
And start a page at which more environment variables can also be
documented.

This allows any page of the docs to link to an env var like this:

```
See the :envvar:`CLIENT_URL` env var.
```

This will:

1. Render the envvar mention as a `<code class="std-envvar">`
   (among other CSS classes), this is then styled specially by the
   Sphinx theme

2. Automatically hyperlink it to the env var's definition on the
   env vars page

3. Generate a warning if an envvar with that name isn't defined
   (if the -n arg to sphinx-build is used)

I think this will also allow docs in the client repo to link to env vars
defined in the h repo docs in this way too, because we have intersphinx
set up (there's already a couple of CLIENT_URL references in the client
docs).

In the future we might be able to use Sphinx's autodoc extension to
generate this envvars page from docstrings in the code.