Use flag service for annotation flag formatter
Now that we have a service to query for flag state, we can add a method
to it to list subsets of flagged annotations and simplify the
`AnnotationFlagFormatter`, removing its knowledge of the session and the
model logic.

To make it clear that the behaviour of the formatter hasn't changed,
I've left the tests as they are, and introduced a test dependency on the
`FlagService`. This might be something to simplify in a later commit,
faking out the `FlagService`, or it may be worth leaving in place to
function as an integration test.