Shortcut empty queries on the `all_flagged` method
We *can* do this because we know that the only subset of an empty set of
annotation IDs is an empty set of annotation IDs, so we don't need to go
talk to the database.

We *should* do this because SQLAlchemy produces potentially inefficient
queries when given `in_` queries with zero-length sequences[1].

This case is already covered in the unit tests, so we know this hasn't
changed the result of the method.

[1]: http://docs.sqlalchemy.org/en/rel_1_1/faq/sqlexpressions.html?highlight=in_#why-does-col-in-produce-col-col-why-not-1-0