Change annotation formatters to accept a resource
Some formatters (including the forthcoming one to add flag counts for
moderators) are going to need to make decisions based on the current
user's permissions, so needs access to the resource object.

This makes some of the tests for the existing formatter slightly less
neat, as they need to wrap the annotation up in something that quacks
like a resource. I also needed to add a new test to the presenter, as
changing the interface didn't end up breaking any of the existing
tests.