Upgrade bleach and html5lib (#4490)
html5lib before 0.99999999 (eight nines) has a couple of XSS
vulnerabilities (CVE-2016-9909 and CVE-2016-9910).

Upgrading to the latest html5lib (0.999999999, nine nines) requires that
we upgrade bleach to 2.0.0, which in turn requires a few small changes
in our code that integrates with bleach. These include:

- attribute filter callables now always take the tag name as the first
  parameter
- linkify callbacks now expose attributes as a dictionary of
  _namespaced_ attributes, so we have to use tuples as keys into the
  dictionary rather than simple strings

At the same time, it made sense to use the Cleaner and LinkifyFilter
classes, new in 2.0, which make it possible to configure sanitisation
and linkification once and reuse the configuration for each string to be
sanitised.