Move memex.links to h.services.links
Moves the links service into the h package alongside other services.
Happily, this services is relatively self-contained, and so this is a
straight move from one place to another.

The only slight "gotcha" is that h.services must now be included before
h.links, otherwise the "add_annotation_link_generator" configuration
directive won't be defined.

This is part of a project to reintegrate the `memex` package into the
main `h` package: https://notes.wtk.io/2017/03/08/reintegrating-memex.