Properly reindex annotations when un-/nipsa'ing a user
We used to go directly to the ElasticSearch index and add or remove the
`"nipsa": true` item to or from the annotation object.

Now that the nipsa flag either gets set based on the user nipsa flag, or
the annotation nipsa flag we have to properly re-index the users'
annotations so we go through the normal search indexing code path.

Considering that most users probably won't have anywhere close to >10k
annotations, I've decided to just load all the annotation ids of a user
in one query.