Fix Docker builds reporting as ".dirty" when they're not
The builds that we push to Docker Hub are built from a `git archive`
tarball, and are thus guaranteed to be built from a clean committed
tree. In Skyliner, however, the whole repository is added to the Docker
image, including the .git directory.

At the moment, all Skyliner builds are reporting themselves as having
dirty trees, because some aspect of the filesystem inside the container
leads git to believe that every single file has been modified. This
seems to be because the git index (`.git/index`) has stale data in its
stat cache, and updating this explicitly with `git update-index
--refresh` or by running `git status` fixes the issue.

In order to be able to do this the `.git` directory needs to be
writeable by the user that runs the application, so this commit:

- ensures that `.git`, if present, is writeable by the `hypothesis` user

- explicitly runs `git update-index -q --refresh` before checking the
  dirty state of the working tree