Add annotation object to transform event
This is done in preparation for the annotation moderation event
subscriber which will need access to the annotation object and not just
to the presented dictionary. This in turn is needed so we can optimise
the database queries for reindexing all annotations by checking if an
annotation is hidden by loading the `Annotation.moderation`
relationship, and thus being able to eager load this data.