Extra most common formatter setup into a fixture
This also means I could get rid of a couple of the formatters where
`has_permission=None` (which was initially to avoid having to pass in
the extra parameter where it would never be used, so its value didn't
matter). Given the value doesn't matter, it makes just as much sense to
pass in a permission-granting value as an exploding one.