Add support for falling through to web service
When the request contains the header key/value
`Hypothesis-Test-Migration: 1`, the fallback upstream will always be we
web service itself. This is to aid testing the migration of the
marketing website (WordPress) to the `web.hypothes.is` subdomain.

To aid this change we also get rid of the warning message response when
no fallback upstream URI is configured. In that case we just send the
request to the web service as well and let it respond with a 404 Not
Found response.