Conditionally redirect users from homepage
Once our marketing website has moved fully off the main application
domain, we'll serve redirects from the root URL that depend on the
user's logged-in status.

If a user is logged in, they are redirected to their own profile page.
If they're not, they are redirected to the main activity stream, unless
another URL is configured using the `HOMEPAGE_REDIRECT_URL` environment
variable, in which case they are sent there.

This change is feature-flagged (`homepage_redirects`).