Index the first of an annotation's references (#4532)
This is the first step toward indexing a top-level annotation's thread
IDs (all annotations which cite the annotation as their top-level
reference, so the entire tree of replies) into Elasticsearch, so that we
don't end up hiding an entire thread when its root annotation gets
hidden by a moderator.

A few notes on the expression for this index:

    [sa.text('("references"[1])')]

This is fairly heavily strewn with punctuation, and I feel I should
explain what I understand of why. The double quotes around "references"
are reasonably understandable: "references", as well as being the name
of the column, is a SQL keyword used when defining foreign keys, so it
needs quoting as a column name. The extra parentheses I understand a
little less well, but without them the migration fails with a syntax
error:

    sqlalchemy.exc.ProgrammingError: (psycopg2.ProgrammingError) syntax error at or near ")"
    LINE 1: ...Y ix__annotation_thread_root ON annotation ("references"[1])
                                                                          ^
     [SQL: 'CREATE INDEX CONCURRENTLY ix__annotation_thread_root ON annotation ("references"[1])']

As for the "1" in the middle: Postgres uses 1-based indexing for its
arrays, rather than 0-based indexing like most languages. Usually,
SQLAlchemy's ORM hides this difference away from us, but we're not in
the ORM world here.