Add a `thread` relationship to annotations
This relationship applies to all top-level annotations, and includes all
the annotations which cite this annotation as the first of their
references (i.e. all annotations for which this annotation is the root
of their thread tree).

Since we started indexing the first element of the `references` array
(see commit 10ba0b33bd), we can look this relationship up efficiently.

We would ideally like to have a reverse `thread_root` relationship here
(for the sake of completeness, rather than because we necessarily want
to preload anything yet), but implementing that relationship in this way
causes SQLAlchemy to generate syntactically invalid SQL queries, along
the lines of:

    ... WHERE ARRAY['9c88f6cc-24eb-11e7-b4ce-fbb096f808ff'::uuid][1] = annotation.id

rather than passing in the single UUID required. We haven't yet managed
to get to the bottom of why it does this, and since we don't need the
`thread_root` relationship yet, we've punted on figuring it out. There's
a good chance that the solution involves either a bug in our custom UUID
type or delving into the murky world of custom comparators[1] -- either
way, that would entail quite a significant amount of work for a feature
we don't yet need.

[1]: http://docs.sqlalchemy.org/en/rel_1_1/orm/extensions/hybrid.html#building-custom-comparators