Add a `thread_ids` field to the search index
This is defined as a `string` type in the index, because Elasticsearch
doesn't appear to make the distinction between single-value and
multi-valued fields. So in practice we're always expecting this to be
provided as a list of strings, hence the plural field name.

I've put this in as a separate commit so that we can theoretically
separate out the migration to add the new field to the mapping from the
task to populate it by reindexing all the annotations. In practice,
though, I think we'll be ok doing these both in a single deployment:
because of the way we've got Elasticsearch configured, it will accept
any fields it doesn't recognise[1], so we don't have to worry about the
indexer trying and failing to index annotations with the old mapping.

[1]: https://hypothes-is.slack.com/archives/C1MA4E9B9/p1492696573403426?thread_ts=1492696027.204889&cid=C1MA4E9B9