Remove fallback and enable redirects to new site
This commit removes the nginx configuration which allows for a
"fallback" origin. This means that instead of sending non-whitelisted
URLs to WordPress (as configured in production), they will now be
handled by the application.

This has the effect of exposing the various redirects which have been
prepared in the application and will "launch" the new blog and marketing
website when released.

It also means that 404s will now be handled by the application.