When indexing a reply, index its top-level parent (#4545)
Now that we're indexing all of the annotation IDs in a thread using the
`thread_ids` field, we want to update this when a new reply gets
created, especially as we're hoping in future to use this field to avoid
a second Elasticsearch query to find reply IDs.

This feels like it might be giving the `add_annotation` task quite a few
distinct responsibilities (it needs to be aware of the multiple indices
used during reindexing, and also requires some knowledge of the
annotation model. As the code currently stands, though, it was a choice
between putting this code directly into the API layer or putting it into
this task -- all the code in-between only has access to the annotation's
ID -- and this is much more a search engine concern than an API concern.