Add strict-origin-when-cross-origin for browsers that support it
Apparently[1] Referrer-Policy allows multiple comma-separated values in
order to allow the rollout of newer policy types while providing a
fallback for older browsers.

As a result, we can add the "strict-origin-when-cross-origin" policy
(which will never send any referrer information to insecure origins) to
take advantage of it as support is rolled out.

[1]: https://www.w3.org/TR/referrer-policy/#unknown-policy-values