Filter out the content of hidden annotations
When an annotation has been hidden by a moderator, but is still being
sent to the client because it has replies, we want to filter out the
annotation content (its text and tags) for anyone but its author (who
doesn't know it's been hidden) and moderators (who get to see
everything).

To avoid having to give the formatter any knowledge of how our
permission model works, the formatter is passed a callable which takes a
group and returns a Boolean value saying whether or not the current user
is a moderator for that group. The behaviour of this callable is the one
bit of functionality (as far as I'm aware) that isn't currently covered
by tests -- I'm open to suggestions on how to cover this without some
seriously convoluted mock-wrangling.