Don't return same publisher group twice
Don't return the same publisher group twice in the /api/profile
response.

Steps to reproduce:

1. Follow the instructions at https://github.com/hypothesis/publisher-account-test-site
   to get the publisher account site setup
2. Login to the publisher test site as "admin" (the publisher group
   creator)

Result: You'll see the same Partner.org publisher group listed twice in
the groups dropdown menu in the client.

Expected: You should only see the group once.

The problem appears to be that in the `_current_groups()` function the
publisher group appears in both the `authority_groups` and the
`_user_groups()`. The admin user, who is the creator of the publisher
group, is also a member of the group.