Show all error messages for JSON validation errors
Previously we used the jsonschema's best_match[1] which tries to choose
the most important error message and returns only that one in the error
response. This results in a sub-optimal experience when trying to use
the API, as sometimes a developer needs to call the API endpoint
multiple times until the server accepts a valid request payload.

This commit changes this behaviour and returns all errors from an
invalid payload as a comma-separated string.

Fixes #4551.

[1]: https://python-jsonschema.readthedocs.io/en/latest/errors/#best-match-and-relevance