Combine unique username and email errors
For a better user experience, we should return both errors in the same
response if both the username and the email is not unique.
We don't add the mismatched authority error to this, because we don't
want to leak any information of existing users back to the client for a
different authority then the one of the authenticated client.